Break down components, simplify interactions
