+++ src/custom/components/swap/EthFlow/pure/WrappingPreview/WrapCard.cosmos.tsx
@@ -1,0 +1,11 @@
+import { WrapCard } from './WrapCard'
+
+import { WRAPPED_NATIVE_CURRENCY as WETH } from 'constants/tokens'
+import { SupportedChainId } from 'constants/chains'
+import { CurrencyAmount } from '@uniswap/sdk-core'
+
+const WrappedEther = WETH[SupportedChainId.MAINNET]
+const wrapAmount = CurrencyAmount.fromRawAmount(WrappedEther, 5.21456 * 10 ** 18)
+const balance = CurrencyAmount.fromRawAmount(WrappedEther, 15.12123 * 10 ** 18)
+
+export default <WrapCard currency={WrappedEther} symbol="ETH" amountToWrap={wrapAmount} balance={balance} />
