+++ src/custom/components/swap/EthFlow/pure/WrappingPreview/WrapCard.tsx
@@ -1,0 +1,46 @@
+import { CurrencyAmount, Currency } from '@uniswap/sdk-core'
+import styled from 'styled-components/macro'
+import CurrencyLogo from 'components/CurrencyLogo'
+import { formatSmart } from 'utils/format'
+import { AMOUNT_PRECISION } from 'constants/index'
+import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
+import { CHAIN_INFO } from 'constants/chainInfo'
+import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
+
+import * as styledEl from './styled'
+
+const BackupTokenImg = styled.img.attrs((attrs) => ({ ...attrs, width: '24px' }))`
+  filter: invert(1);
+`
+
+interface WrapCardProps {
+  symbol: string
+  balance?: CurrencyAmount<Currency>
+  amountToWrap?: CurrencyAmount<Currency>
+  currency: Currency
+  chainId?: number
+}
+
+export function WrapCard(props: WrapCardProps) {
+  const { symbol, balance, amountToWrap, currency, chainId } = props
+  const hasLogoUri = currency.isNative || Boolean(currency instanceof WrappedTokenInfo && currency.logoURI)
+
+  return (
+    <styledEl.WrapCardWrapper>
+      {/* logo */}
+      {!hasLogoUri && chainId && currency.equals(WRAPPED_NATIVE_CURRENCY[chainId]) ? (
+        <BackupTokenImg alt="token-img" src={CHAIN_INFO[chainId].logoUrl} />
+      ) : (
+        <CurrencyLogo currency={currency} size="24px" />
+      )}
+      {/* amount to wrap/unwrap */}
+      <styledEl.BalanceLabel>
+        <strong>
+          {formatSmart(amountToWrap, AMOUNT_PRECISION) || '-'} {symbol}
+        </strong>
+      </styledEl.BalanceLabel>
+      {/* user balance */}
+      <styledEl.BalanceLabel>Balance: {formatSmart(balance, AMOUNT_PRECISION) || '-'}</styledEl.BalanceLabel>
+    </styledEl.WrapCardWrapper>
+  )
+}
