+++ src/custom/components/swap/EthFlow/pure/WrappingPreview/WrappingPreview.cosmos.tsx
@@ -1,0 +1,10 @@
+import { WrappingPreview } from '.'
+
+import { wrappingPreviewProps } from '../../mocks'
+
+const Fixtures = {
+  withBalance: <WrappingPreview {...wrappingPreviewProps} />,
+  unknonwnBalance: <WrappingPreview {...wrappingPreviewProps} nativeBalance={undefined} wrappedBalance={undefined} />,
+}
+
+export default Fixtures
