+++ src/custom/components/swap/EthFlow/pure/WrappingPreview/index.tsx
@@ -1,0 +1,1 @@
+export { WrappingPreview, type WrappingPreviewProps } from './WrappingPreview'
